backend: refactor api/detect

+++ backend/detect/views.py
@@ -12,12 +12,14 @@ def detection(request):
         name = request.POST.get("name")
         result = request.POST.get("result")
         detail = request.POST.get("detail")
-        pic_at_rest = request.FILES.get("pic_at_rest")
-        pic_forehead_wrinkle = request.FILES.get("pic_forehead_wrinkle")
-        pic_eye_closure = request.FILES.get("pic_eye_closure")
-        pic_smile = request.FILES.get("pic_smile")
-        pic_snarl = request.FILES.get("pic_snarl")
-        pic_lip_pucker = request.FILES.get("pic_lip_pucker")
+        # pic_at_rest = request.FILES.get("pic_at_rest")
+        # pic_forehead_wrinkle = request.FILES.get("pic_forehead_wrinkle")
+        # pic_eye_closure = request.FILES.get("pic_eye_closure")
+        # pic_smile = request.FILES.get("pic_smile")
+        # pic_snarl = request.FILES.get("pic_snarl")
+        # pic_lip_pucker = request.FILES.get("pic_lip_pucker")
+        
+        print(name + " is uploading detection result")
 
         if not User.objects.filter(name=name).exists():
             return JsonResponse({
@@ -31,29 +33,29 @@ def detection(request):
 
         save_path = f"media/{name}_{save_name}/"
 
-        with open(save_path + "pic_at_rest.jpg", "wb") as f:
-            for chunk in pic_at_rest.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_at_rest.jpg", "wb") as f:
+        #     for chunk in pic_at_rest.chunks():
+        #         f.write(chunk)
                 
-        with open(save_path + "pic_forehead_wrinkle.jpg", "wb") as f:
-            for chunk in pic_forehead_wrinkle.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_forehead_wrinkle.jpg", "wb") as f:
+        #     for chunk in pic_forehead_wrinkle.chunks():
+        #         f.write(chunk)
                 
-        with open(save_path + "pic_eye_closure.jpg", "wb") as f:
-            for chunk in pic_eye_closure.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_eye_closure.jpg", "wb") as f:
+        #     for chunk in pic_eye_closure.chunks():
+        #         f.write(chunk)
                 
-        with open(save_path + "pic_smile.jpg", "wb") as f:
-            for chunk in pic_smile.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_smile.jpg", "wb") as f:
+        #     for chunk in pic_smile.chunks():
+        #         f.write(chunk)
                 
-        with open(save_path + "pic_snarl.jpg", "wb") as f:
-            for chunk in pic_snarl.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_snarl.jpg", "wb") as f:
+        #     for chunk in pic_snarl.chunks():
+        #         f.write(chunk)
                 
-        with open(save_path + "pic_lip_pucker.jpg", "wb") as f:
-            for chunk in pic_lip_pucker.chunks():
-                f.write(chunk)
+        # with open(save_path + "pic_lip_pucker.jpg", "wb") as f:
+        #     for chunk in pic_lip_pucker.chunks():
+        #         f.write(chunk)
 
         record = Result(name=name, result=0, detail="检测中", comment="", save_path=save_path, time=current_time)
         record.save()
@@ -85,8 +87,9 @@ def detection(request):
 def history(request):
     if request.method == "POST":
         name = request.POST.get("name")
-        print(name)
         page = int(request.POST.get("page", 1))
+        
+        print(name + " is requesting history for page " + str(page))
 
         if not User.objects.filter(name=name).exists():
             return JsonResponse({
@@ -122,6 +125,8 @@ def upload_comment(request):
         id = request.POST.get("id")
         comment = request.POST.get("comment")
         
+        print("Commenting on result " + id)
+        
         if not Result.objects.filter(id=id).exists():
             return JsonResponse({
                 "code": 400,
@@ -145,8 +150,14 @@ def upload_comment(request):
 
 def clear(request):
     if request.method == "POST":
-        Result.objects.all().delete()
+        secret_key = request.POST.get("secret_key")
+        if secret_key == "123456":
+            Result.objects.all().delete()
+            return JsonResponse({
+                "code": 200,
+                "message": "Clear success"
+            })
         return JsonResponse({
-            "code": 200,
-            "message": "Clear success"
+            "code": 400,
+            "message": "Invalid secret key"
         })
